Implement a simple EventLoop mechanism for use with basic C#-based implementations of CommonScript.
This event loop is optional and serves as both a one-size-fits-most implementation and also a documentation of a practical implementation, as certain contexts will likely manage things differently.

There still needs to be a better multi-threading wake-up mechanism that isn't just a CPU spin. Right now, there's just a 1ms sleep in the while loop.