Simplfy the configuration for JavaScript runtime.
The value converter available to extension methods should be bound to the current runtime instance and not require the implementer to pass in task references.
Running a task should run as a promise and not require the user to implement task result status interpretation themselves.
The extension for stdout could be simplified further with a direct listener rather than requiring the user to add extensions based on the magic string 'io_stdout'.
Default implementations for sleep and delay_invoke are provided. This only makes sense in languages/platforms that don't have de-facto async behavior for these.